AP-484: ensure HLS audio streams are TS packetized (#5)
we were seeing an issue where the new native HLS playback in Chrome and Edge 141+ would
prevent playback of audio streams. Pacifica was the notable example to the contrary as
it worked in all browsers.

this also cleans up a deprecated comment that related to a no-longer extant/supported
server-wide configuration file.